Reorganize extension.py into a directory to prepare for new API
