+++ asdf/extension/__init__.py
@@ -1,0 +1,20 @@
+"""
+Support for plugins that extend asdf to serialize
+additional custom types.
+"""
+from ._legacy import (
+    AsdfExtension,
+    AsdfExtensionList,
+    BuiltinExtension,
+    default_extensions,
+    get_default_resolver,
+)
+
+
+__all__ = [
+    "AsdfExtension",
+    "AsdfExtensionList",
+    "BuiltinExtension",
+    "default_extensions",
+    "get_default_resolver",
+]
