+++ asdf/util.py
@@ -269,6 +269,14 @@ def resolve_name(name):
     return ret
 
 
+# Kludge to cover up the fact that BuiltinExtension was moved from extension.py
+# to extension/_legacy.py.  Can be removed once BuiltinExtension is dropped
+# in asdf 3.0.
+_CLASS_NAME_OVERRIDES = {
+    "asdf.extension._legacy.BuiltinExtension": "asdf.extension.BuiltinExtension",
+}
+
+
 def get_class_name(obj, instance=True):
     """
     Given a class or instance of a class, returns a string representing the
@@ -282,9 +290,9 @@ def get_class_name(obj, instance=True):
     instance: bool
         Indicates whether given object is an instance of the class to be named
     """
-
     typ = type(obj) if instance else obj
-    return "{}.{}".format(typ.__module__, typ.__name__)
+    class_name = "{}.{}".format(typ.__module__, typ.__name__)
+    return _CLASS_NAME_OVERRIDES.get(class_name, class_name)
 
 
 def minversion(module, version, inclusive=True, version_path='__version__'):
